+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.2.0
+
+-   Change accent spec to match themes.
+
+-   Add more variants:
+
+    -   `OKSolar Red`
+    -   `OKSolar Green`
+    -   `OKSolar Purple`
+    -   `OKSolar Purple`
+    -   `OKSolar Blue`
+
 ## v0.1.0
 
 -   Initial release